refactor: unify duplicate StellarImpactCards components across all use-case sections
- Create shared StellarImpactCards component at use-cases/shared/ that
  supports both 'detailed' (expandable cards) and 'simple' variants
- Extract per-section card data into co-located data.tsx files
- Replace 5 duplicated ~400-line implementations with thin wrappers
  that import from the shared component and pass local data
- Preserve all animations, hover effects, and grid layouts
- No TypeScript errors — build verified successfully
