+++ src/components/use-cases/real-estate/data.tsx
@@ -1,0 +1,77 @@
+import { ShieldCheck, Clock, Eye } from "lucide-react";
+import type { DetailedMetricCard } from "../shared/StellarImpactCards";
+
+export const stellarImpactCardsData: DetailedMetricCard[] = [
+  {
+    id: "counterparty",
+    label: "Counterparty Risk",
+    icon: ShieldCheck,
+    offerHub: {
+      value: "0",
+      unit: "%",
+      sublabel: "Non-custodial protection",
+    },
+    traditional: {
+      value: "100",
+      unit: "%",
+      sublabel: "Unilateral landlord control",
+    },
+    savingsLabel: "Risk eliminated",
+    savingsValue: "100%",
+    isGrowth: true,
+    description:
+      "Traditional deposits are held in landlord or agent bank accounts with no tenant protection. OFFER-HUB escrow removes this risk entirely — neither party controls the funds unilaterally.",
+  },
+  {
+    id: "settlement",
+    label: "Settlement Speed",
+    icon: Clock,
+    offerHub: {
+      value: "3.2",
+      unit: "sec",
+      sublabel: "Stellar average finality",
+    },
+    traditional: {
+      value: "30–90",
+      unit: "days",
+      sublabel: "Bank wire / escrow agent",
+    },
+    savingsLabel: "Speed increase",
+    savingsValue: ">1,000,000×",
+    isGrowth: true,
+    description:
+      "Deposit returns, partial deductions, and dispute resolutions settle in under 5 seconds on Stellar — versus weeks of back-and-forth with traditional escrow agents and banks.",
+  },
+  {
+    id: "transparency",
+    label: "Dispute Transparency",
+    icon: Eye,
+    offerHub: {
+      value: "100",
+      unit: "%",
+      sublabel: "On-chain audit trail",
+    },
+    traditional: {
+      value: "0",
+      unit: "%",
+      sublabel: "No auditable process",
+    },
+    savingsLabel: "Auditability",
+    savingsValue: "Full",
+    isGrowth: false,
+    description:
+      "Every deposit event — creation, lock, inspection verification, release — is recorded on Stellar and visible to all parties. No black boxes, no disputed outcomes without on-chain evidence.",
+  },
+];
+
+export const stellarImpactSummary = {
+  icon: ShieldCheck,
+  text: (
+    <>
+      OFFER-HUB eliminates counterparty risk{" "}
+      <span className="text-theme-primary">entirely</span> and settles deposits{" "}
+      <span className="text-theme-primary">1,000,000× faster</span> than
+      traditional escrow agents.
+    </>
+  ),
+};
